requirements.txt
streamlit==1.28.0
google-generativeai==0.3.0
langchain==0.0.350
langchain-community==0.0.12
langchain-google-genai==0.0.8
chromadb==0.4.18
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
faiss-cpu==1.7.4
transformers==4.35.0
torch==2.1.0
datasets==2.14.6
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.8.2
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
wordcloud==1.9.2
python-dotenv==1.0.0
streamlit-authenticator==0.2.3
streamlit-chat==0.1.1
streamlit-option-menu==0.3.6
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
beautifulsoup4==4.12.2
requests==2.31.0
asyncio==3.4.3
aiohttp==3.9.1
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
elasticsearch==8.11.0
pinecone-client==2.2.4
weaviate-client==3.25.3
openai==1.3.0
cohere==4.37
anthropic==0.7.7
huggingface-hub==0.19.4
accelerate==0.24.1
bitsandbytes==0.41.3
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2
optuna==3.4.0
mlflow==2.8.1
wandb==0.16.0
evidently==0.4.11
prometheus-client==0.19.0
psutil==5.9.6
memory-profiler==0.61.0
line-profiler==4.1.1
py-spy==0.3.14
locust==2.17.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0